+++ app/core/permission_value.py
@@ -1,0 +1,6 @@
+class PermissionValue:
+    practice_area_lead = "PracticeAreaAdmin"
+    global_admin = "GlobalAdmin"
+    project_team_member = "ProjectTeamMember"
+    project_admin = "ProjectAdmin"
+    self_value = "Self"
